+++ src/plugin.ts
@@ -21,7 +21,7 @@ import {
     PluginContext,
 } from "./types/plugin";
 import { dedent } from "./util/dedent";
-import { ExecutableGraph } from "./util/graph/executable";
+import { ExecutableGraph } from "./util/graph/executable-graph";
 import { commandToDot, graphToDot } from "./util/graph/visualisation/dot";
 import { HELP } from "./util/help";
 import { LOG, Level } from "./util/logging";
